LayoutTests/imported/w3c/web-platform-tests/cookies/meta-blocked-expected.txt


PASS Cookie is not set from `<meta>`.
